1.改用第三方二维码软件，改进二维码扫描效率。2.游戏改为两副牌

+++ core/FiveTenKGame/src/com/uc/fivetenkgame/player/Player.java
@@ -284,9 +284,9 @@ public class Player implements PlayerContext {
 		viewController.setCards(mPlayerModel.getCardList());
 		// สฃำเลฦสý
 		List<Integer> leftCardsNum = new ArrayList<Integer>();
-		leftCardsNum.add(new Integer(18));
-		leftCardsNum.add(new Integer(18));
-		leftCardsNum.add(new Integer(18));
+		leftCardsNum.add(new Integer(36));
+		leftCardsNum.add(new Integer(36));
+		leftCardsNum.add(new Integer(36));
 		viewController.setCardNumber(leftCardsNum);
 	}
